[WIP] Factored out utility functions, and am adding more to support more decompression and compilation options
